+++ WolkConnect/ReadingType.py
